Fix default session timeout
Due to a bug in Pax Web, the default session timeout for Opencast was
set to 4 minutes instead of 4 hours. This caused users to be logged out
automatically if they do not have an additional remember-me cookie (e.g.
LTI users don't have that). This basically means that users are unable
to watch videos longer than four minutes for which they need
authentication.

This is a temporary fix for #6260