Don't run attach transcription if AmberScript transcription failed
If an AmberScript transcription job fails – for example if there is no
speech in the video – the service will recognize the error, write that
to the database, but then still continue with trying to attach the
workflow, even overwriting the error state written to the database
again.

This is due to the code just executing the remainder of the code used
for processing successful transcriptions. Instead, it should immediately
continue with the next transcription job.