Document OpenSearch plugin requirements
This documentation patch moves the plugin and configuration requirements
of OpenSearch into the central OpenSearch guide instead of scattering
the information across multiple guides. They all link the main guide
for configuration anyway.

This also switches from documenting Elasticsearch to documenting
OpenSearch in a few places since that is what we suggest to use after
all.