Fix jQuery mime type
Due to recent changes related to the graphQL patches, jQuery was served
as `application/octet-stream` which could lead to browsers no longer
loading the library. This broke the REST docs. Just putting the library
into the `scripts` folder fixes the issue since it's one of the
hard-coded locations where mime type mapping is enabled. The previous
location `/js/` was no such location.

This fixes #6383